// front-end/src/module/model/server/instituicao-curso-ocorrencia-periodo-professor-periodo-aula.ts
export class InstituicaoCursoOcorrenciaPeriodoProfessorPeriodoAula {
    id?: string;
    inicio: string;
    fim: string;
    dia: DayOfWeek;
    label: string;
}

export enum DayOfWeek {
    Sunday = 0,
        Monday = 1,
        Tuesday = 2,
        Wednesday = 3,
        Thursday = 4,
        Friday = 5,
        Saturday = 6,
}